% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carregar_banco_de_dados.R
\name{carregar_banco}
\alias{carregar_banco}
\title{Carregar banco de dados}
\usage{
carregar_banco(banco_de_dados)
}
\arguments{
\item{banco_de_dados}{= 'exemplo.xlsx'}
}
\description{
Entrada deve ser um banco de dados xlsx do medplus.
Essa funcao retira as 3 primeiras linhas e atualiza o nome das colunas
}
