% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agendamento_limpeza.R
\name{limpeza_agendamentos}
\alias{limpeza_agendamentos}
\title{Limpeza para dados do relatorio agendamentos - mantendo apenas os dados validos}
\usage{
limpeza_agendamentos(dados)
}
\arguments{
\item{Dataframe}{}
}
\description{
Essa funcao funciona para os dados do relatorio agendamentos e retira os dados que poluem o banco de dados.
}
\details{
Plano = 'Convenio Cortesia C/ Cobranca de Honorarios' sao retirados

Procedimentos do tipo 'A- REPETICAO', 'GRATUIDADE', 'Retorno' sao retirados

Procedimentos do tipo 'Bloqueado' sao substituidos para 'Atendidos'

Pacientes com o nome 'Nao agendar' ou 'Teste' sao retirados

E substituido o NomeStatus = 'Chegou', 'Na Espera', 'Sem Status', 'Em Consulta' para 'Atendido'

E removido o NomeStatus = 'Confirmado'
}
