% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procedimento_limpeza.R
\name{limpeza_procedimentos}
\alias{limpeza_procedimentos}
\title{Limpeza para dados do relatorio procedimentos - mantendo apenas os dados validos}
\usage{
limpeza_procedimentos(dados)
}
\arguments{
\item{Dataframe}{}
}
\description{
Essa funcao funciona para os dados do relatorio procedimentos e retira os dados que poluem o banco de dados.
}
\details{
Dados do tipo 'A- REPETICAO', 'GRATUIDADE', 'Retorno' sao retirados

Plano = 'Convenio Cortesia C/ Cobranca de Honorarios' sao retirados

Mantem apenas os StatusAgen: 'Atendido' e 'Cumprido e Recebido'
}
