% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padronizar_data.R
\name{padronizar_data}
\alias{padronizar_data}
\title{Arrumar formato da data}
\usage{
padronizar_data(vetor_data)
}
\arguments{
\item{vetor_data}{= 'DD/MM/AAAA 00:00:00' string}
}
\description{
Entrada deve ser um vetor data no formado DD/MM/AAAA 00:00:00.
Essa funcao arruma o padrao data do medplus para o padrao data do r.
}
