% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padronizar_valor.R
\name{padronizar_valor}
\alias{padronizar_valor}
\title{Padroniza o valor do DF}
\usage{
padronizar_valor(vetor)
}
\arguments{
\item{vetor}{do tipo string = 200,000 | 200,00 | 200,0}
}
\value{
vetor numerico = 200
}
\description{
Entrada deve ser um vetor de valores.
Essa funcao substitui virgula por ponto e transforma em numerico.
}
